+++ src/lib/vault/index.ts
@@ -5,6 +5,7 @@ export * from "./cross-tab-sync";
 export * from "./discovery";
 export * from "./graph";
 export * from "./hub-discovery";
+export * from "./live-note";
 export * from "./map-earned";
 export * from "./neighborhood";
 export * from "./note-query";